디테일 화면 1차 완성
