+++ app/src/main/res/drawable/rounded_image.xml
@@ -1,0 +1,8 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <corners android:radius="80dp" />
+    <solid android:color="#FFB600" />
+    <stroke
+        android:width="5dp"
+        android:color="#FF7900" />
+</shape>
