+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,13 @@
     <string name="my_page_introduction_tag">소개</string>
     <string name="my_page_introduction">안녕하세요.^^\n저는 개발자 홍길동 입니다.\n만나서 반갑습니다.\n앞으로도 잘 부탁 드립니다.</string>
     <string name="my_page_img_tag">이미지</string>
+
+    <string name="name">배기정</string>
+    <string name="phone_number">전화번호</string>
+    <string name="phone_number_data">010-1234-5678</string>
+    <string name="email">이메일</string>
+    <string name="email_data">bgj@mail.com</string>
+    <string name="team">소속팀</string>
+    <string name="team_data">내배캠 안드로이드 3기 태조</string>
+    <string name="memo">메모</string>
 </resources>
